refactor: 중복되는 테스트 코드를 테스트 픽스처 공유 클래스화
