+++ src/test/java/boardSetting/PieceTest/CarTest.java
@@ -1,28 +1,23 @@
 package boardSetting.PieceTest;
 
+import boardSetting.TestFIxture;
 import domain.position.Position;
 import domain.Team;
-import domain.piece.Blank;
 import domain.piece.Car;
-import domain.piece.Piece;
-import domain.PieceProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.*;
 
 public class CarTest {
 
     private Car car;
-    private TestPieceProvider testBoard;
+    private TestFIxture testBoard;
 
     @BeforeEach
     public void setUp() {
         car = new Car(Team.CHO);
-        testBoard = new TestPieceProvider();
+        testBoard = new TestFIxture();
     }
 
     @Test
@@ -46,34 +41,5 @@ public class CarTest {
         boolean isCarMove = car.canMove(currentPosition, targetPosition, testBoard);
 
         assertThat(isCarMove).isFalse();
-    }
-
-
-    private static class TestPieceProvider implements PieceProvider {
-        private final Map<Position, Boolean> boardState = new HashMap<>();
-        private boolean defaultState = true;
-
-        void setBlank(Position pos) {
-            boardState.put(pos, false);
-        }
-
-        void setAllBlank() {
-            this.defaultState = true;
-        }
-
-        @Override
-        public boolean isBlank(Position position) {
-            return boardState.getOrDefault(position, defaultState);
-        }
-
-        @Override
-        public boolean isCannon(Position position) {
-            return false;
-        }
-
-        @Override
-        public Piece getPiece(Position position) {
-            return new Blank();
-        }
     }
 }
