+++ src/test/java/boardSetting/PieceTest/ElephantTest.java
@@ -1,26 +1,23 @@
 package boardSetting.PieceTest;
 
-import domain.PieceProvider;
+import boardSetting.TestFIxture;
 import domain.position.Position;
 import domain.Team;
 import domain.piece.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ElephantTest {
 
-    private TestPieceProvider testBoard;
+    private TestFIxture testBoard;
     private Elephant elephant;
 
     @BeforeEach
     void setUp() {
         elephant = new Elephant(Team.CHO);
-        testBoard = new TestPieceProvider();
+        testBoard = new TestFIxture();
     }
 
     @Test
@@ -43,33 +40,5 @@ public class ElephantTest {
 
         boolean isCanMove = elephant.canMove(currentPosition, targetPosition, testBoard);
         assertThat(isCanMove).isFalse();
-    }
-
-    private static class TestPieceProvider implements PieceProvider {
-        private final Map<Position, Boolean> boardState = new HashMap<>();
-        private boolean defaultState = true;
-
-        void setBlank(Position pos) {
-            boardState.put(pos, false);
-        }
-
-        void setAllBlank() {
-            this.defaultState = true;
-        }
-
-        @Override
-        public boolean isBlank(Position position) {
-            return boardState.getOrDefault(position, defaultState);
-        }
-
-        @Override
-        public boolean isCannon(Position position) {
-            return false;
-        }
-
-        @Override
-        public Piece getPiece(Position position) {
-            return new Blank();
-        }
     }
 }
